pyproject.toml
[project]
name = "watsonx-mcp"
version = "0.1.0"
description = "Watsonx MCP server (Matrix-ready, SSE, local process)"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
  { name = "Ruslan Magana Vsevolodovna", email = "contact@ruslanmv.com" }
]
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
]

# Use the official MCP SDK plus Starlette/Uvicorn for the /health route
dependencies = [
  "mcp>=1.8.0",
  "starlette>=0.37",
  "uvicorn>=0.30",
  "ibm-watsonx-ai>=1.1.15",
  "python-dotenv>=1.0.1",
]

[tool.setuptools]
packages = ["watsonx_mcp"]
package-dir = {"" = "."}

[tool.pytest.ini_options]
addopts = "-q"
